Switch default
It's probably better to have the locmem backend by default -- actually,
it's probably better to have the request backend, but I don't want that
when in unit tests.

We need to figure out how to override the test environment set up so we
always use the locmem backend, in the same way Django does with the
email backend.